requirements.txt
numpy
pandas
tifffile
Pillow
scipy
